Migrate AppComponent and Application classes to Hilt
