+++ WooCommerce/src/main/kotlin/com/woocommerce/android/WooCommerce.kt
@@ -14,8 +14,6 @@ import com.google.android.gms.common.GoogleApiAvailability
 import com.woocommerce.android.analytics.AnalyticsTracker
 import com.woocommerce.android.analytics.AnalyticsTracker.Stat
 import com.woocommerce.android.cardreader.CardReaderManager
-import com.woocommerce.android.di.AppComponent
-import com.woocommerce.android.di.DaggerAppComponent
 import com.woocommerce.android.di.WooCommerceGlideModule
 import com.woocommerce.android.network.ConnectionChangeReceiver
 import com.woocommerce.android.push.FCMRegistrationIntentService
@@ -39,6 +37,7 @@ import dagger.MembersInjector
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
 import dagger.android.HasAndroidInjector
+import dagger.hilt.android.HiltAndroidApp
 import org.greenrobot.eventbus.Subscribe
 import org.greenrobot.eventbus.ThreadMode
 import org.wordpress.android.fluxc.Dispatcher
@@ -84,12 +83,6 @@ open class WooCommerce : MultiDexApplication(), HasAndroidInjector, ApplicationL
 
     private var connectionReceiverRegistered = false
 
-    open val component: AppComponent by lazy {
-        DaggerAppComponent.builder()
-                .application(this)
-                .build()
-    }
-
     companion object {
         private const val SECONDS_BETWEEN_SITE_UPDATE = 60 * 60 // 1 hour
     }
@@ -111,11 +104,6 @@ open class WooCommerce : MultiDexApplication(), HasAndroidInjector, ApplicationL
     override fun onCreate() {
         super.onCreate()
 
-        // enable strict mode in debug builds
-        if (BuildConfig.DEBUG) {
-            enableStrictMode()
-        }
-
         // Disables Volley debug logging on release build and prevents the "Marker added to finished log" crash
         // https://github.com/woocommerce/woocommerce-android/issues/817
         if (!BuildConfig.DEBUG) {
@@ -124,8 +112,6 @@ open class WooCommerce : MultiDexApplication(), HasAndroidInjector, ApplicationL
 
         val wellSqlConfig = WooWellSqlConfig(applicationContext)
         WellSql.init(wellSqlConfig)
-
-        component.inject(this)
 
         FeedbackPrefs.init(this)
 
@@ -159,37 +145,6 @@ open class WooCommerce : MultiDexApplication(), HasAndroidInjector, ApplicationL
                 this, BuildConfig.ZENDESK_DOMAIN, BuildConfig.ZENDESK_APP_ID,
                 BuildConfig.ZENDESK_OAUTH_CLIENT_ID
         )
-    }
-
-    /**
-     * enables "strict mode" for testing - should NEVER be used in release builds
-     */
-    private fun enableStrictMode() {
-        // return if the build is not a debug build
-        if (!BuildConfig.DEBUG) {
-            WooLog.e(T.UTILS, "You should not call enableStrictMode() on a non debug build")
-            return
-        }
-
-        StrictMode.setThreadPolicy(
-            ThreadPolicy.Builder()
-                .detectDiskWrites()
-                .detectNetwork()
-                .penaltyLog()
-                .penaltyFlashScreen()
-                .build()
-        )
-
-        StrictMode.setVmPolicy(
-            VmPolicy.Builder()
-                .detectActivityLeaks()
-                .detectLeakedSqlLiteObjects()
-                .detectLeakedClosableObjects()
-                .detectLeakedRegistrationObjects()
-                .penaltyLog()
-                .build()
-        )
-        WooLog.w(T.UTILS, "Strict mode enabled")
     }
 
     override fun onAppComesFromBackground() {
