--- WooCommerce/src/main/kotlin/com/woocommerce/android/di/AppComponent.kt
@@ -1,52 +1,0 @@
-package com.woocommerce.android.di
-
-import android.app.Application
-import com.woocommerce.android.WooCommerce
-import com.woocommerce.android.media.ProductImagesServiceModule
-import com.woocommerce.android.push.FCMServiceModule
-import com.woocommerce.android.ui.login.LoginAnalyticsModule
-import com.woocommerce.android.util.crashlogging.CrashLoggingModule
-import dagger.BindsInstance
-import dagger.Component
-import dagger.android.AndroidInjectionModule
-import dagger.android.AndroidInjector
-import org.wordpress.android.fluxc.module.ReleaseNetworkModule
-import org.wordpress.android.fluxc.module.ReleaseOkHttpClientModule
-import org.wordpress.android.login.di.LoginServiceModule
-import javax.inject.Singleton
-
-@Singleton
-@Component(
-    modules = [
-        AndroidInjectionModule::class,
-        ApplicationModule::class,
-        AppConfigModule::class,
-        ReleaseNetworkModule::class,
-        ReleaseOkHttpClientModule::class,
-        SelectedSiteModule::class,
-        ThreadModule::class,
-        ActivityBindingModule::class,
-        FCMServiceModule::class,
-        LoginAnalyticsModule::class,
-        LoginServiceModule::class,
-        NetworkStatusModule::class,
-        CurrencyModule::class,
-        ProductImagesServiceModule::class,
-        SupportModule::class,
-        OrderFetcherModule::class,
-        CrashLoggingModule::class
-    ]
-)
-interface AppComponent : AndroidInjector<WooCommerce> {
-    override fun inject(app: WooCommerce)
-
-    // Allows us to inject the application without having to instantiate any modules, and provides the Application
-    // in the app graph
-    @Component.Builder
-    interface Builder {
-        @BindsInstance
-        fun application(application: Application): Builder
-
-        fun build(): AppComponent
-    }
-}
