+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/AppConfigModule.kt
@@ -5,10 +5,13 @@ import com.woocommerce.android.AppPrefs
 import com.woocommerce.android.BuildConfig
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import org.wordpress.android.fluxc.network.UserAgent
 import org.wordpress.android.fluxc.network.rest.wpcom.auth.AppSecrets
 import java.util.Locale
 
+@InstallIn(SingletonComponent::class)
 @Module
 class AppConfigModule {
     companion object {
