+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/CurrencyModule.kt
@@ -4,9 +4,12 @@ import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.util.CurrencyFormatter
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import org.wordpress.android.fluxc.store.WooCommerceStore
 import javax.inject.Singleton
 
+@InstallIn(SingletonComponent::class)
 @Module
 class CurrencyModule {
     @Singleton
