+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/NetworkStatusModule.kt
@@ -4,8 +4,11 @@ import android.content.Context
 import com.woocommerce.android.tools.NetworkStatus
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
+@InstallIn(SingletonComponent::class)
 @Module
 class NetworkStatusModule {
     @Provides
