+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/SelectedSiteModule.kt
@@ -4,9 +4,12 @@ import android.content.Context
 import com.woocommerce.android.tools.SelectedSite
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import org.wordpress.android.fluxc.store.SiteStore
 import javax.inject.Singleton
 
+@InstallIn(SingletonComponent::class)
 @Module
 class SelectedSiteModule {
     @Provides
