+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/SupportModule.kt
@@ -5,10 +5,13 @@ import com.woocommerce.android.support.SupportHelper
 import com.woocommerce.android.support.ZendeskHelper
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import org.wordpress.android.fluxc.store.AccountStore
 import org.wordpress.android.fluxc.store.SiteStore
 import javax.inject.Singleton
 
+@InstallIn(SingletonComponent::class)
 @Module
 class SupportModule {
     @Singleton
