+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/ThreadModule.kt
@@ -3,8 +3,11 @@ package com.woocommerce.android.di
 import com.woocommerce.android.util.CoroutineDispatchers
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import org.wordpress.android.util.helpers.Debouncer
 
+@InstallIn(SingletonComponent::class)
 @Module
 class ThreadModule {
     @Provides
