+++ WooCommerce/src/main/kotlin/com/woocommerce/android/media/ProductImagesServiceModule.kt
@@ -2,7 +2,10 @@ package com.woocommerce.android.media
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 
+@InstallIn(SingletonComponent::class)
 @Module
 internal abstract class ProductImagesServiceModule {
     @ContributesAndroidInjector
