+++ WooCommerce/src/main/kotlin/com/woocommerce/android/push/FCMServiceModule.kt
@@ -2,7 +2,10 @@ package com.woocommerce.android.push
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 
+@InstallIn(SingletonComponent::class)
 @Module
 abstract class FCMServiceModule {
     @ContributesAndroidInjector
