+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/login/LoginAnalyticsModule.kt
@@ -2,10 +2,13 @@ package com.woocommerce.android.ui.login
 
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import org.wordpress.android.fluxc.store.AccountStore
 import org.wordpress.android.fluxc.store.SiteStore
 import org.wordpress.android.login.LoginAnalyticsListener
 
+@InstallIn(SingletonComponent::class)
 @Module
 class LoginAnalyticsModule {
     @Provides
