+++ WooCommerce/src/main/kotlin/com/woocommerce/android/util/crashlogging/CrashLoggingModule.kt
@@ -9,8 +9,11 @@ import com.woocommerce.android.util.locale.LocaleProvider
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
+@InstallIn(SingletonComponent::class)
 @Module
 abstract class CrashLoggingModule {
     companion object {
