+++ WooCommerce/src/release/kotlin/com/woocommerce/android/di/FluxCModule.kt
@@ -1,0 +1,18 @@
+package com.woocommerce.android.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import org.wordpress.android.fluxc.module.ReleaseBaseModule
+import org.wordpress.android.fluxc.module.ReleaseNetworkModule
+import org.wordpress.android.fluxc.module.ReleaseOkHttpClientModule
+import org.wordpress.android.fluxc.module.ReleaseWCNetworkModule
+
+@InstallIn(SingletonComponent::class)
+@Module(
+    includes = [
+        ReleaseNetworkModule::class,
+        ReleaseOkHttpClientModule::class
+    ]
+)
+abstract class FluxCModule
